flashcards/requirements.txt
Flask>=3.0.0,<4.0.0
PyYAML>=6.0.0,<7.0.0
pytest>=8.0.0,<9.0.0
